+++ src/DepthCameraSensor.cc
@@ -336,6 +336,8 @@ bool DepthCameraSensor::CreateCamera()
   this->dataPtr->depthCamera->SetImageHeight(height);
   this->dataPtr->depthCamera->SetNearClipPlane(near);
   this->dataPtr->depthCamera->SetFarClipPlane(far);
+  this->dataPtr->depthCamera->SetVisibilityMask(
+      cameraSdf->VisibilityMask());
 
   this->AddSensor(this->dataPtr->depthCamera);